+++ tests/libs/estdlib/test_epmd.erl
@@ -1,0 +1,91 @@
+%
+% This file is part of AtomVM.
+%
+% Copyright 2024 Paul Guyot <pguyot@kallisys.net>
+%
+% Licensed under the Apache License, Version 2.0 (the "License");
+% you may not use this file except in compliance with the License.
+% You may obtain a copy of the License at
+%
+%    http://www.apache.org/licenses/LICENSE-2.0
+%
+% Unless required by applicable law or agreed to in writing, software
+% distributed under the License is distributed on an "AS IS" BASIS,
+% WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+% See the License for the specific language governing permissions and
+% limitations under the License.
+%
+% SPDX-License-Identifier: Apache-2.0 OR LGPL-2.1-or-later
+%
+
+-module(test_epmd).
+
+-export([test/0]).
+
+test() ->
+    ok = test_client(),
+    ok = test_two_clients(),
+    ok.
+
+test_client() ->
+    {ok, Pid1} = erl_epmd:start_link(),
+    noport = erl_epmd:port_please("test_epmd", "host.invalid"),
+    noport = erl_epmd:port_please("test_epmd", "localhost"),
+    {ok, Creation1} = erl_epmd:register_node("test_epmd", 12345),
+    {port, 12345, Version} = erl_epmd:port_please("test_epmd", "localhost"),
+    6 = Version,
+    {error, already_registered} = erl_epmd:register_node("test_epmd", 12345),
+    {error, already_registered} = erl_epmd:register_node("test_epmd_new", 12346),
+    {ok, Names} = erl_epmd:names("localhost"),
+    true = lists:member({"test_epmd", 12345}, Names),
+    MonitorRef1 = monitor(process, Pid1),
+    unlink(Pid1),
+    erl_epmd:stop(),
+    shutdown =
+        receive
+            {'DOWN', MonitorRef1, process, Pid1, Reason1} -> Reason1
+        after 5000 -> timeout
+        end,
+
+    {ok, Pid2} = erl_epmd:start_link(),
+    noport = erl_epmd:port_please("test_epmd", "localhost"),
+    {ok, Creation2} = erl_epmd:register_node("test_epmd", 12345),
+    true = Creation1 =/= Creation2,
+    MonitorRef2 = monitor(process, Pid2),
+    unlink(Pid2),
+    erl_epmd:stop(),
+    shutdown =
+        receive
+            {'DOWN', MonitorRef2, process, Pid2, Reason2} -> Reason2
+        after 5000 -> timeout
+        end,
+
+    ok.
+
+test_two_clients() ->
+    {ok, Pid1} = erl_epmd:start_link(),
+    {ok, _Creation1} = erl_epmd:register_node("test_epmd_1", 12345),
+    unregister(erl_epmd),
+    {ok, Pid2} = erl_epmd:start_link(),
+    {ok, _Creation2} = erl_epmd:register_node("test_epmd_2", 12346),
+    {ok, Names} = erl_epmd:names("localhost"),
+    true = lists:member({"test_epmd_1", 12345}, Names),
+    true = lists:member({"test_epmd_2", 12346}, Names),
+    unlink(Pid2),
+    MonitorRef2 = monitor(process, Pid2),
+    erl_epmd:stop(),
+    shutdown =
+        receive
+            {'DOWN', MonitorRef2, process, Pid2, Reason2} -> Reason2
+        after 5000 -> timeout
+        end,
+    register(erl_epmd, Pid1),
+    MonitorRef1 = monitor(process, Pid1),
+    unlink(Pid1),
+    erl_epmd:stop(),
+    shutdown =
+        receive
+            {'DOWN', MonitorRef1, process, Pid1, Reason1} -> Reason1
+        after 5000 -> timeout
+        end,
+    ok.